Move writers into members_serialize
